enemies rotate in place when not moving
made it so that when enemies that follow patterns or move randomly, or that roam, rotate in place when they aren't moving, just to give them a little life so they aren't completely still in between moves. And this is if they even have any down time, they can be set to never stop moving and not pause between movements.