implemented enemy sprites
Implemented the sprite and animation for the level 2 enemy to the enemy that moves randomly. Also for some reason the 1st enemy had its sprite removed and animation like undone or something somehow so I had to redo that. Jacob should take a look at it if he sees this, looks a little jittery this time